Split StorageSlot into StorageSlotTransient
