+++ contracts/utils/README.adoc
@@ -34,7 +34,8 @@ Miscellaneous contracts and libraries containing utility functions you can use t
  * {Strings}: Common operations for strings formatting.
  * {ShortString}: Library to encode (and decode) short strings into (or from) a single bytes32 slot for optimizing costs. Short strings are limited to 31 characters.
  * {SlotDerivation}: Methods for deriving storage slot from ERC-7201 namespaces as well as from constructions such as mapping and arrays.
- * {StorageSlot}: Methods for accessing specific storage slots formatted as common primitive types. Also include primitives for reading from and writing to transient storage (only value types are currently supported).
+ * {StorageSlot}: Methods for accessing specific storage slots formatted as common primitive types. 
+ * {StorageSlotTransient}: primitives for reading from and writing to transient storage (only value types are currently supported).
  * {Multicall}: Abstract contract with a utility to allow batching together multiple calls in a single transaction. Useful for allowing EOAs to perform multiple operations at once.
  * {Context}: A utility for abstracting the sender and calldata in the current execution context.
  * {Packing}: A library for packing and unpacking multiple values into bytes32
@@ -129,6 +130,8 @@ Ethereum contracts have no native concept of an interface, so applications must 
 {{SlotDerivation}}
 
 {{StorageSlot}}
+
+{{StorageSlotTransient}}
 
 {{Multicall}}
