+++ scripts/generate/templates/StorageSlotMock.js
@@ -2,7 +2,7 @@ const format = require('../format-lines');
 const { TYPES } = require('./Slot.opts');
 
 const header = `\
-pragma solidity ^0.8.24;
+pragma solidity ^0.8.20;
 
 import {Multicall} from "../utils/Multicall.sol";
 import {StorageSlot} from "../utils/StorageSlot.sol";
@@ -40,18 +40,6 @@ function get${name}Storage(uint256 key) public view returns (${type} memory) {
 }
 `;
 
-const transient = ({ type, name }) => `\
-event ${name}Value(bytes32 slot, ${type} value);
-
-function tload${name}(bytes32 slot) public {
-    emit ${name}Value(slot, slot.as${name}().tload());
-}
-
-function tstore(bytes32 slot, ${type} value) public {
-    slot.as${name}().tstore(value);
-}
-`;
-
 // GENERATE
 module.exports = format(
   header,
@@ -63,7 +51,6 @@ module.exports = format(
       TYPES.filter(type => type.isValueType).map(type => storageSetValueType(type)),
       TYPES.filter(type => type.isValueType).map(type => storageGetValueType(type)),
       TYPES.filter(type => !type.isValueType).map(type => storageSetNonValueType(type)),
-      TYPES.filter(type => type.isValueType).map(type => transient(type)),
     ),
   ).trimEnd(),
   '}',
